feat: ship kiosk standalone mode and production hardening
