+++ apps/frontend-admin/src/app/kiosk/page.tsx
@@ -1,0 +1,5 @@
+import { KioskShell } from '@/components/kiosk/kiosk-shell'
+
+export default function KioskPage() {
+  return <KioskShell />
+}
