+++ apps/frontend-admin/src/app/layout.tsx
@@ -15,7 +15,7 @@ import './globals.css'
 const DRAWER_ID = 'app-drawer'
 
 /** Routes that render without the app shell (navbar, sidebar) */
-const BARE_ROUTES = ['/login']
+const BARE_ROUTES = ['/login', '/kiosk']
 
 export default function RootLayout({
   children,
