+++ apps/frontend-admin/src/lib/websocket.ts
@@ -11,6 +11,7 @@ class WebSocketManager {
     this.socket = io(socketUrl, {
       transports: ['websocket', 'polling'],
       autoConnect: true,
+      withCredentials: true,
       reconnection: true,
       reconnectionDelay: 1000,
       reconnectionAttempts: 5,
